@@ -1,15 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
+﻿using System.Windows.Controls;
 
 namespace caMonPageMod.HMIDisp233.No2
 {
@@ -18,7 +7,7 @@ namespace caMonPageMod.HMIDisp233.No2
 	/// </summary>
 	public partial class Type0 : UserControl
 	{
-		public Type0(DataClass dc)
+		public Type0(HoanLampStates dc)
 		{
 			DataContext = dc;
 			InitializeComponent();
