@@ -7,7 +7,7 @@ namespace caMonPageMod.HMIDisp233.No2
 	/// </summary>
 	public partial class Type1k : UserControl
 	{
-		public Type1k(DataClass dc)
+		public Type1k(HoanLampStates dc)
 		{
 			DataContext = dc;
 			InitializeComponent();
